chore: update component name to be clearer
