--- apps/hubble-stats/components/NetworkPoolTable/index.ts
@@ -1,1 +1,0 @@
-export { default as NetworkPoolTable } from './NetworkPoolTable';
