--- apps/hubble-stats/components/NetworkTokenTable/index.ts
@@ -1,1 +1,0 @@
-export { default as NetworkTokenTable } from './NetworkTokenTable';
