--- apps/hubble-stats/containers/NetworkTablesContainer/index.ts
@@ -1,1 +1,0 @@
-export { default as NetworkTablesContainer } from './NetworkTablesContainer';
