--- apps/hubble-stats/containers/PoolChartsContainer/index.ts
@@ -1,1 +1,0 @@
-export { default as PoolChartsContainer } from './PoolChartsContainer';
