+++ apps/hubble-stats/containers/PoolOverviewChartsContainer/index.ts
@@ -1,0 +1,1 @@
+export { default as PoolOverviewChartsContainer } from './PoolOverviewChartsContainer';
