+++ apps/hubble-stats/containers/PoolOverviewTableContainer/index.ts
@@ -1,0 +1,1 @@
+export { default as PoolOverviewTableContainer } from './PoolOverviewTableContainer';
