+++ apps/hubble-stats/containers/PoolWrappingTableContainer/index.ts
@@ -1,0 +1,1 @@
+export { default as PoolWrappingTableContainer } from './PoolWrappingTableContainer';
