+++ apps/hubble-stats/containers/index.ts
@@ -1,12 +1,13 @@
 export { KeyMetricsTableContainer } from './KeyMetricsTableContainer';
-export { NetworkTablesContainer } from './NetworkTablesContainer';
 export { OverviewChartsContainer } from './OverviewChartsContainer';
 export { OverviewChipsContainer } from './OverviewChipsContainer';
-export { PoolChartsContainer } from './PoolChartsContainer';
+export { PoolOverviewChartsContainer } from './PoolOverviewChartsContainer';
 export { PoolMetadataTableContainer } from './PoolMetadataTableContainer';
 export { PoolOverviewCardContainer } from './PoolOverviewCardContainer';
+export { PoolOverviewTableContainer } from './PoolOverviewTableContainer';
 export { PoolTransactionsTableContainer } from './PoolTransactionsTableContainer';
 export { PoolWrappingChartsContainer } from './PoolWrappingChartsContainer';
+export { PoolWrappingTableContainer } from './PoolWrappingTableContainer';
 export { ShieldedTablesContainer } from './ShieldedTablesContainer';
 
 export { Layout } from './Layout';
