+++ apps/hubble-stats/data/index.ts
@@ -1,5 +1,4 @@
 export { default as getKeyMetricsData } from './getKeyMetricsData';
-export { default as getNetworkTablesData } from './getNetworkTablesData';
 export {
   default as getOverviewChartsData,
   type OverviewChartsDataType,
@@ -12,6 +11,10 @@ export {
 export { default as getPoolMetadataData } from './getPoolMetadataData';
 export { default as getPoolOverviewCardData } from './getPoolOverviewCardData';
 export {
+  default as getPoolOverviewTableData,
+  type PoolOverviewTableDataType,
+} from './getPoolOverviewTableData';
+export {
   default as getPoolTransactionsData,
   type PoolTransactionDataType,
 } from './getPoolTransactionsData';
@@ -19,4 +22,8 @@ export {
   default as getPoolWrappingChartsData,
   type PoolWrappingChartsDataType,
 } from './getPoolWrappingChartsData';
+export {
+  default as getPoolWrappingTableData,
+  type PoolWrappingTableDataType,
+} from './getPoolWrappingTableData';
 export { default as getShieldedTablesData } from './getShieldedTablesData';
